Started database addition
Added DBDemo.java to project structure. This will be used as a base
for the DatabaseManager class in our project. In addition, I made some
slight modifications to the README.md file, which will tell how to compile
and run our project from the command line. It is important to note that all
csv files should be in the top level directory when using the provided
commands. In addition, the working directory should also be the top level
directory of our project.